+++ include/AcceleratorPedal.hpp
@@ -6,20 +6,20 @@ void fallingAcceleratorPWMPulse();
 void risingAcceleratorPWMPulse();
 
 // TODO: Adjust Values
-const int RAW_FULL_ACCELERATOR = 2020;
-const int RAW_NO_ACCELERATOR = 1500;
-const int RAW_BRAKE = 980;
+const short int RAW_FULL_ACCELERATOR = 2020;
+const short int RAW_NO_ACCELERATOR = 1500;
+const short int RAW_BRAKE = 980;
 
-int ACCELERATOR_PWM_PIN_INPUT;
-int ACCELERATOR_PWM_PIN_OUTPUT;
+short int ACCELERATOR_PWM_PIN_INPUT;
+short int ACCELERATOR_PWM_PIN_OUTPUT;
 int acceleratorPreviousTime;
-int acceleratorInterruptPin;
-int rawAcceleratorInputPWM;
+byte acceleratorInterruptPin;
+byte rawAcceleratorInputPWM;
 
 bool acceleratorIsManual;
 
-int acceleratorPercent;
-int manualPWMInput;
+byte acceleratorPercent;
+byte manualPWMInput;
 
 void fallingAcceleratorPWMPulse();
 void risingAcceleratorPWMPulse();
